+++ R/python-packages.R
@@ -36,13 +36,38 @@
 #'   will instead look at all loaded packages and discover their associated
 #'   Python requirements.
 #'
+#' @param force Boolean; force configuration of the Python environment? Note
+#'   that `configure_environment()` is a no-op within non-interactive \R
+#'   sessions. Use this if you require automatic environment configuration, e.g.
+#'   when testing a package on a continuous integration service.
+#'
 #' @export
-configure_environment <- function(package = NULL) {
+configure_environment <- function(package = NULL, force = FALSE) {
   
+  # no-op when Python has not yet been initialized
   if (!is_python_initialized())
     return(FALSE)
   
-  # find Python requirements  
+  # allow opt-out through envvar
+  auto <- Sys.getenv("RETICULATE_AUTOCONFIGURE", unset = "TRUE")
+  if (auto %in% c("FALSE", "False", "0"))
+    return(FALSE)
+  
+  # disallow in non-interactive R sessions unless forced
+  # (even if force is set, do not allow unless user has explicitly
+  # promised they're not on CRAN)
+  ok <- interactive() || (force && identical(Sys.getenv("NOT_CRAN"), "true"))
+  if (!ok)
+    return(FALSE)
+  
+  # disallow environment configuration when not using a Python environment
+  config <- py_config()
+  root <- dirname(dirname(config$python))
+  ok <- is_virtualenv(root) || is_condaenv(root)
+  if (!ok)
+    return(FALSE)
+  
+  # find Python requirements
   reqs <- python_package_requirements(package)
   if (length(reqs) == 0)
     return(FALSE)
@@ -112,7 +137,7 @@ configure_environment <- function(package = NULL) {
       
       # read installed packages lazily
       if (is.null(pip_installed_packages)) {
-        pip_installed_packages <- pip_freeze(python = python)
+        pip_installed_packages <- pip_freeze(python = config$python)
       }
       
       # construct requirement string
@@ -131,7 +156,7 @@ configure_environment <- function(package = NULL) {
     } else {
       
       # read installed packages lazily
-      envpath <- dirname(dirname(python))
+      envpath <- dirname(dirname(config$python))
       conda <- miniconda_conda()
       
       if (is.null(conda_installed_packages)) {